Integrate API & Unit Test
Implement a service that will make an HTTP request to the following API endpoint: https://randomuser.me/api/?results=10. This API returns a list of random users.
Connect the service you created to your application. Retrieve the list of users from the API and use it to replace the hard coded list from step 3.

- Add unit testing where required